A modification needed on (some perhaps) 64 bit processor implementations

+++ types.h
@@ -1,9 +1,7 @@
-/*
- * $Header: /home/sater/bridge/bigdeal/RCS/types.h,v 1.2 1999/12/11 09:42:54 sater Exp $
- */
+#include <stdint.h>
 
 typedef unsigned char byte;
-typedef unsigned long dword;
+typedef uint32_t dword;			/* not unsigned long, because of 64 bit words implementations */
 
 #define RMDsize 160
 #define RMDbytes (RMDsize/8)
